Restructure, add FreeMonthsForm
