+++ src/features/tools/campaign-codes/index.tsx
@@ -1,7 +1,7 @@
 import { CampaignFilter } from 'api/generated/graphql'
-import { CampaignCodeFilter } from 'features/tools/campaign-codes/CampaignCodeFilter'
-import { CampaignCodeTable } from 'features/tools/campaign-codes/CampaignCodeTable'
-import { CreateCampaignCode } from 'features/tools/campaign-codes/CreateCampaignCode'
+import { CampaignCodeFilter } from 'features/tools/campaign-codes/components/CampaignCodeFilter'
+import { CampaignCodeTable } from 'features/tools/campaign-codes/components/CampaignCodeTable'
+import { CreateCampaignCode } from 'features/tools/campaign-codes/components/CreateCampaignCode'
 import { initialCampaignFilter } from 'features/tools/campaign-codes/utils'
 import { Card, CardsWrapper } from 'hedvig-ui/card'
 import { MainHeadline } from 'hedvig-ui/typography'
